Corrected the atan2 function
the atan2 function gave totally wrong results...

It has been corrected and seems to be fine now
(changes... in other than the first quadrant wrong values of pi have been
added)

Wrote a arcus cosine function

that function similar to the atan function is build upon a lookuptable.
But here the table is a balanced binary tree in an array.
I extracted to signifiand and exponent of the input floating point numbers

and converted them to 0.[], well only when the absolute value of the float was
smaller are equals one. Than that number (unsigned char) was compared to
vales of that binary tree, to find the right index of values of a second
binary tree. That second tree had values of a first order taylor polynome
calcution which gave the end result.

1. Own extraction of the important bits of the float.
2. BBTree maximum 4 nodes searched (and used unsigned char)
3. First order taylor polynom.

--> 3 times the speed for Arduino UNO, maximum error under 3 degrees,
average error maximum 0.4 degrees

Advice only use it, when you realy need the speed!